New proto parser and filtering.
This is working toward the modularization of Wire. We want three steps:

 * Parsing
 * Filtering
 * Emitting

This commit is the first draft of the first two steps.

`WireParser` is the public API that the command line and any build system or plugins
will interface with to parse and filter the protos. They will hand the result off to
a the yet-unwritten Wire code gen or a code generator of their own choosing.