Fork MessageWriter to use JavaPoet.
This creates a new class TypeWriter that is mostly a conversion
of MessageWriter, but it uses linked Wire types instead of the
unlinked protoparser types.

This is successfully passing 12 of the 35 WireCompiler tests
with slight alterations to imports (JavaPoet doesn't yet support
static imports.)

The options and extensions are still to be completed. That work
should be straightforward, but I'm saving it for a follow up.
Hopefully with that change I can switch the default to be the
new JavaPoet backend.

Also add a dependency on Guava (we should have had this long ago)
and also on the snapshot of JavaPoet (which we need to include
documentation on enum constants).