Switch to JavaPoet.
This large change flips a flag to turn on the new code path, and it changes
some insignificant formatting in many of the wire compiler tests.

Most of the compiler tests are changed due to things like static
imports, which aren't yet supported in JavaPoet. When those sorts
of things are supported, the code formatting will be improved.