Introduce MessageAdapter for reading and writing instances.
This exposes a public API for an internal concept which is a class instance that is responsible for reading and writing protos of a specific type. This allows consumers to cache these instances for use while serializing to/from bytes. With this change all other means of reading and writing encoded data have been removed except for `toByteArray()` on instances.

Additionally, a few types and a lot of methods were removed from being exposed the public API as they are implementation details of serialization.