New Schema and Location models.
The new Schema class allows Linker and Pruner to be hidden as implementation
details of the wire-schema module. Other users of the wire-schema library
shouldn't need to use these directly.

The new Location class locates a message, enum, extension or service within
a .proto file. Later we can add support for fields, options, and RPCs, which
will allow for better error messages when a link cannot be resolved.