Better error messages when linking fails.
Also move some type validation from ProtoParser into the schema package.
In particular, we no longer enforce that an RPC's type is not a scalar
in the protoparser layer; that happens in the linker.

In the future we can move more validation into the linker (unique tag
names, etc.) which should improve usability.