Use WireType instead of Message.Datatype.
The nice upside of this change is that we now include the protobuf type name
like squareup.protos.person.Person in the annotation where it is known.

The other nice impact of this is we don't have the weird pseudotypes ENUM
and MESSAGE that don't correspond to anything on the wire.

The drawback is that the generated @ProtoField enum now uses a string instead
of an enum constant for known scalars.