Optimize list movement for repeated fields.
We incorrectly mixed mutable (copy constructor) and immutable (default constructor) list values when initializing the builder. In the mutable case, we did an entire copy of the immutable list only to make another immutable copy when building. In the immutable case, we allowed it to pass through when building as an optimization. A mutable list is always used in the builder for consistency, but with some special behavior for efficiency:

* When the copy constructor is used, retain the immutable instance inside a mutable container from the original value and re-use it if it goes unchanged when building.
* When the default constructor is used, wrap an emptyList() in a mutable container and use it if no values were added when building.
* If the mutable container is mutated in any way, make a copy of the immutable data and into a proper mutable list and allow mutation.