Improve performance of TagMap toString, expose helper API.
Calling `extensions(true)` and `get(Extension)` are expensive in terms of computation and allocation. This changes the `TagMap` to do an effectively-single-pass traversal of the underlying array appending to an existing `StringBuilder`. This method will also be used by generated code, and the implementation of `toString` in the message adapter has been updated to reflect more closely the behavior of what generated code would look like.