Prune options.
This got a little out of hand. In order to correctly prune options we need
to have a proper definition of what the options are, and for that we need
to consult Google's descriptor.proto, particularly for things like deprecated
and java_package, otherwise those get list.

And then descriptor.proto isn't quite complete, it doesn't model the field
default option. It's special and goes directly on the field.

Pruning the options themselves isn't so bad, but it means we need to call
retainAll() on a bunch of things we weren't previously.

This also fixes pruning of oneofs, previously they were missing.