Simplify the identifier set and marking mechanism.
This changes the roles of 'includes' and 'excludes', and makes marking run
in three phases rather than two.

In the first phase we walk through the entire schema, looking for stuff to
include. These are the roots. This will allow us to support wildcards in
includes in a follow up PR.

In the second phase we follow transitive dependencies from the roots and
mark those.

In the third phase we retain only what is marked.

This is different from what we had before because the IdentifierSet and
MarkSet are now dumber. I deleted the tests, though new tests might come
back once the interactions get more interesting. Tests that were lost won't
make sense when we have wildcards, since IdentifierSet will need more
policy on what an include or exclude means.