Do not throw when parsing a null repeated field
MessageTypeAdapter throws an IllegalArgumentException when attempting to parse a repeated field that is set to null in a JSON string.  Ideally, the null field would never have been included in the JSON to begin with, but the source may not be as smart as Wire.  MessageTypeAdapter should handle it anyway.