Begin to parse build.wire configuration files.
This extracts the general-purpose parts of the parser out of ProtoParser
and into a standalone file, Parser.java.

It then uses that file to build a new parser for build.wire files, which
currently only support type configs. The type configs do not yet support
any deep validation; that will come in a follow up that will hook these
up to SchemaLoader.