Honor profiles for Wire configuration.
Previously we used command line parameters like --proto-adapter. With this
change we instead discover android.wire or java.wire files and use those
to customize the generated .java code.

This renames SchemaBuilder to RepoBuilder. The new name is preferred because
it now builds both Schema objects and also Profile objects. The class now
represents a source tree.

This renames wire-schema-tests to wire-tests. Much of the movement in this
PR is to handle this move.

This also moves some JavaGenerator tests from wire-java-generator to wire-tests.
The reason is that I want to reuse the new RepoBuilder type.