Write a new .proto file loader that will work as WireRun prescribes
It's more flexible in how inputs are provided - fully-qualified paths,
directories, or .zips. And it validates that the parsed paths are in
the expected directories.

The Root class keeps track of a user-provided directory where .proto
files are loaded.

The NewSchemaLoader loads the user's provided sources, and the protos
they import.