Lazily load from the proto path.
Previously the proto path and source path were symmetric: all types and members and
options were always loaded, parsed, and linked.

With this change we lazily load, parse, and link from the proto path.

The biggest fallout of this change is that some things are newly null when
pruning including the options map and field types. This should be cleaned up
in a follow-up PR, but this PR is big enough as-is.