Add validation to check proto3 Enum has always 0 value
As the official Google document says, proto3 Enum must have a zero value. To support this, we'll add a validation for this in EnumType#validate(...). Since we need to know the proto version of ProtoFile we validate, Syntax? should be passed to the function.

Ref: https://developers.google.com/protocol-buffers/docs/proto3#enum
```
There must be a zero value, so that we can use 0 as a numeric default value.
```