Add manifest concept for partitioning Swift output
Because code in a Swift module lives in a flat namespace (as opposed to Java/Kotlin and its packages) we need to be able to partition the messages into logical modules.

This change introduces a YAML manifest with a simple schema to facilitate that. The format is as follows:

    ModuleName:
      dependencies:
        - OtherModule
      include:
        - example.feature.Message
      exclude:
        - example.logging.Settings

    OtherModule:
      include:
        - example.other.Type

All of the maps under the module name are optional.