Optimization: Reduce virtual function references
Resolving the virtual pointers for the generic decode functions has a cost associated with it. By adding non-generic implementations for common types we let the compiler insert a direct function pointer for those types and remove the more expensive runtime lookup.

This reduces the time to do 10k AllTypes round trips from about 1.78s to 1.38s