Optimization: Use raw data buffer for reading Swift proto data
The Swift `Data` object has some overhead when accessing bytes in certain ways. This moves to using an `UnsafePointer<UInt8>` to manage that access more manually and avoid that overhead. It also moves the read progress cursor into a new object (`ReadBuffer`) so that `ProtoReader` doesn't have to manage it manually.

This takes the 10k AllTypes round trip time down from about 1.38s to 1.21s.