Extract schema partitioning by manifest to be language-agnostic
The manifest will partition a schema into a dedicated paritioned schema value type. That value type can then be used with a SchemaHandler factory to create an uber SchemaHandler which will produce the partitioned modules.

Eventually we want to extract this above WireRun and targets, but that will come later.