Maintain a stack of frames for message nesting
This lets us consolidate the message end pointer and the unknown fields data into a single structure, and also gives us a place to store additional info, like whether or not the message currently being decoded is using proto2 or proto3 syntax (this functionality was also added).

This doesn't seem to have any performance implications. The AllTypes round trip test runs at the same speed with and without this change.