Add property to retrieve proto syntax version in Swift
This adds an explicit property to retrieve the proto syntax used for a given message (either proto2 or proto3). We can get the same information by doing something like `value is Proto2Codable`, but currently (Swift 5.2) that does a full dynamic cast under the hood and is quite slow. By using a method on the protocol instead we're just resolving a virtual function pointer.

This has no noticeable performance impact yet since we only currently check the syntax in one rarely-called place, but this sets the stage for adding proto3 encoding in ProtoWriter without having a noticeable performance cost.