Add message stack to Swift ProtoWriter
This lets us track whether or not the current message is a proto2 or proto3 message. We'll use this in the future to determine whether or not to encode default values or to use packed vs. discrete array packing by default.

The performance impact of this change is negligible. It may raise the AllTypes round trip time from about 0.88s to 0.89s, but it's tough to tell since it's within the margin of error.