Change packing default for proto2 vs proto3
This changes it so that the default packing behavior for proto3 messages is to pack them, while leaving the default for proto2 messages as discrete serialization.

This has negligible performance impact, and looks to add about a 1% slowdown for the AllTypes round trip test, possibly raising it from ~0.90s to ~0.91s.