Fallback to stub protoadapters when java.time is missing
This allows clients, specially Android ones, not to be forced to
to address the issue if they don't have to. If a client is
actually using google.protobuf.Duration or g.p.Instant, the
generated code will not compile and clients will be able to
detect the missing parts at compile time, instead of having
the stub proto adapter failing at runtime.