Configuration to turn on and off option annotations
Currently declared options are emitted by default, and applied options
are not. In a follow-up release I'll turn everything on by default.
I've chosen this mode because it's easiest to upgrade. Otherwise you
can't upgrade downstream things without upgrading upstream things
first.

This required some changes to linking because we now need to recover
the package name of the file the option is declared in. Previously this
was not necessary.