Remove prepare command in Wire.podspec
This isn't necessary since we do the same thing in a script phase within the test spec. By removing the prepare command we remove the requirement that any consumer of the pod needs to have Java installed, and that only becomes a requirement if you want to run the tests.