Fix issue where imports fail with an error when building on Windows.
This may not be a complete fix - this was the bare minimum required
to get things to work on Windows for the sample app, and
Linker.normaliseImportPath has a number of flaws.

The issue appears to be that Linker.fileLinkers is used with paths in
both OS-native path format (eg. with \ separators on Windows) and
source file format (eg. with / separators). If these match, there is
no issue, but if they don't match (eg. loading a file with / separators
in imports on Windows), imports aren't resolved correctly.

A more elegant solution may be to use a separator-agnostic type for
the key to Linker.fileLinkers.