Pull reflection out of RuntimeMessageAdapter
Promote RuntimeMessageAdapter from jvmMain to commonMain. Leave the reflective
code behind in a new reflection.java file.

This requires a new interface, MessageBinding, that encapsulates the features
implemented by reflection.

This is in anticipation of a brand new implementation of MessageBinding that
uses a parsed schema to encode and decode values without any code generation
whatsoever.