ReverseProtoWriter: like ProtoWriter but without measurement
In performance measurement I observed Wire's encodedSize() showing up too much in
profiles. This is likely due to the fact that this method must be invoked repeatedly
for deeply-nested datastructures.

Attempt to improve things by changing encoding from front-to-back to back-to-front,
which allows us to capture the data size for free as a side-effect.