Modernize WireTask APIs to use more modern Gradle APIs and be configuration caching friendly
This leans heavily on [managed properties](https://docs.gradle.org/current/userguide/custom_gradle_types.html#managed_properties), proper `Property` types, and fixes `project` accessor issues that break configuration caching in the `WireTask`. It's a breaking change but since 4.0 is in alpha it seemed like now is a good time, and it would be hard to do this without breaking anything.

Breaking in the following ways:
- Binary breaking change (obviously)
- Kotlin source breaking change (also obviously)
- _not_ a Groovy source breaking change (possibly not obvious)