Clean up WirePluginTest and enable configuration caching as part of tests
This ensures individual tests are properly isolated from each other by defining and clearing their project and gradle home dirs between runs. Before, they would reuse caches and possibly return false positives on caches. This also adds `--configuration-cache` to the test arguments to catch any regressions to that support in the future. It does run more slowly, but that's the cost of full isolation.

Note that this also inlines the test runner initializer to the property, because it's needless to put it directly in the setup method here unless you're expecting it to possibly throw an exception and want a cleaner error message.