Modernize WireTask APIs to use more modern Gradle APIs and be configuration caching friendly (#2088)
* Modernize WireTask APIs to use more modern Gradle APIs and be configuration caching friendly

This leans heavily on [managed properties](https://docs.gradle.org/current/userguide/custom_gradle_types.html#managed_properties), proper `Property` types, and fixes `project` accessor issues that break configuration caching in the `WireTask`. It's a breaking change but since 4.0 is in alpha it seemed like now is a good time, and it would be hard to do this without breaking anything.

Breaking in the following ways:
- Binary breaking change (obviously)
- Kotlin source breaking change (also obviously)
- _not_ a Groovy source breaking change (possibly not obvious)

* Clean up WirePluginTest and enable configuration caching as part of tests

This ensures individual tests are properly isolated from each other by defining and clearing their project and gradle home dirs between runs. Before, they would reuse caches and possibly return false positives on caches. This also adds `--configuration-cache` to the test arguments to catch any regressions to that support in the future. It does run more slowly, but that's the cost of full isolation.

Note that this also inlines the test runner initializer to the property, because it's needless to put it directly in the setup method here unless you're expecting it to possibly throw an exception and want a cleaner error message.

* Move back to setup because of tmpFolder 🤡