Fix Swift test build
The build requirements changed in Xcode 12.5 such that we need to include a new framework. This change adds the /usr/lib directory that contains that framework into the build and linker arguments so that the Swift build can find it.

Also fixes a few small things in the GitHub action:
- `macOS-latest` -> `macos-latest`
- Explicitly specifies the Xcode version and path to use