Avoid NPEs with kotlin 1.7.20 (#2310)
With kotlin 1.7.20, some of the source sets (that are nullable properties)
are actually null. The wire gradle plugin code was treating these cases as
exceptional and was throwing exceptions. With this fix, the cases are handled
more gracefully by creating an empty source set and using that instead.

No test is added, since the wire build itself suffers from this problem if
kotlin 1.7.20 is used to build, and so a successful build with 1.7.20 is itself
a test of this code.