Restrict building the Swift WireCompiler to Mac only
Without a platform definition CocoaPods assumes you want to build for macOS, iOS, watchOS, and tvOS. It really only makes sense to run the Wire compiler on macOS, so restrict to that platform. This also makes the builds a little faster and the deployment process easier since you don't need all of the simulators to run the tests.