feat: Adjust Gradle plugin `outputs` to be computed lazily
At Faire, we have Gradle convention plugins in use to share common configuration of our many (~50) protobuf projects. To do this, we leverage a [Gradle Convention plugin](https://docs.gradle.org/current/samples/sample_convention_plugins.html) to wrap our common settings. One issue we have is that we want to make the `out` directory configurable by scripts, but due to how Wire requires computation at wire configuration time, it is not possible.

This PR moves the computation of targets to be wrapped in a `ListProperty<>` and pushes the execution of `Action<>` to be when tasks are configured. The `plugin-playground` project shows a simplified case.