New @WireEnclosingType annotation, so R8 doesn't prune too much (#3123)
* New @WireEnclosingType annotation, so R8 doesn't prune too much

We've observed a surprising behavior where Kotlin reflection
assumes the nested classes of non-Kotlin classes are themselves
non-Kotlin classes.

This assumption is reasonable, but breaks if R8 prunes the
`@Metadata` annotation on an enclosing class. The horrible result
is a nested class that has a `@Metadata` annotation but isn't considered
to be a Kotlin class. Yuck.

This marker annotation should help us to configure R8 to avoid
this unfortunate situation, and it might prove useful for other people
as well.

* Fix a test

* Fix another test

* Spotless