Avoid eagerly resolving the configuration
Calling task.source.isEmpty does some sort of eager resolution of
the configurations that were added to the task source. This in
turn can trigger other actions within Gradle and cause weird
behaviour (in our case, it caused other plugins to get applied
and eventually tripped a mutation guard.

The wire rootset code has a isEmpty flag it keeps which, according
to the comments, is there explicitly to avoid resolving the
configuration, so we can use that instead.