Parse trailers even when an error occurs
It is possible for gRPC error responses to be empty and for servers to
send the error information within trailers without also sending them in
headers. In these cases, Wire will propagate an exception due to no
response body and will not attempt to read trailers, relying only on
headers instead.

This patch allows Wire to attempt to parse the trailers for
non-streaming requests instead of only relying on headers.